Replace message content
Ideas have solidified a bit since it was first written,
it now looks like the best approach is to split as much of the system as is feasible into containers,
leaving pretty much just systemd in the root container,
which pivots into a new mount namespace or mount tree and restarts other services in its container.
If they socket activate, or store their fds in systemd,
then they can update without service interruption.